Поправила тип вводимых данных
Задача №6: Отработка арифметических операций
Написать программу, которая будет выполнять следующие действия:

1. Ввод трех чисел с клавиатуры X,Y,Z

2. Нахождение и вывод на экран среднего арифметического чисел X,Y,Z

3. Деление среднего арифметического на 2 без остатка

4. Если полученное число больше 3, то вывести на экран сообщение "Программа выполнена корректно"

+++ src/HelloWorld.java
@@ -4,11 +4,11 @@
         public static void main(String[] args) {
             Scanner sc = new Scanner(System.in);
             System.out.print("Введите число: ");
-            double x = new Scanner(System.in).nextInt();
+            double x = new Scanner(System.in).nextDouble();
             System.out.print("Введите число: ");
-            double y = new Scanner(System.in).nextInt();
+            double y = new Scanner(System.in).nextDouble();
             System.out.print("Введите число: ");
-            double z = new Scanner(System.in).nextInt();
+            double z = new Scanner(System.in).nextDouble();
             double avsum = (x+y+z)/3;
             System.out.println(avsum);
             int sum = (int)(avsum /= 2);
